Added the disc golf data to the stand-in database
For completeness, added the steps to the Makefile to convert the shape
files to KML, and then take those to the SQLite database.

There are two KML files for the tee locations (the short and long tees),
but the shape files don't have projection files, and they don't line up
with the polyline data for the holes when I use the same projection that
the polyline data for the holes has.  For now, I'm not using those KMLs
at all.  Also, there are two holes for which the short tee polyline is
distinct from the long tee polyline, and I'm simply discarding them for
now since I don't have the short tee info at all.

Also fixed a typo in latitude, which had two the first 't' doubled.